Property Sales in USA
We have been given the historical dataset containing data on Property Sales in the USA. This dataset contains 1460 observations with 18 variables, both integers and categorical values, describe housing conditions and properties and does not contain any missing values. Aim of a project to fit the best model                          
to predict housing prices.